requirements.txt
chromadb==0.5.23
langchain==0.3.13
langchain-community==0.3.13
langchain-openai==0.2.14
openai==1.58.1
PySide6==6.8.1
python-dotenv==1.0.1
